requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues



azure-functions

transformers==4.28.1
pandas==1.3.4
tweepy==4.14.0

-f https://download.pytorch.org/whl/torch_stable.html
torch==1.13.0+cpu
torchvision==0.14.0+cpu

DateTime==5.0
python-dotenv==0.19.2
azure-storage-blob==2.1.0
requests==2.26.0
email-to==0.1.0
numpy==1.24.2

nltk==3.8.1
